Midst of refactoring to class-like behavior
