+++ .vs/ProjectSettings.json
