+++ scraper/webscraper_config.json
@@ -1,0 +1,3 @@
+{
+    "OPEN_AI_API_KEY": "sk-1234567890"
+}
