+++ world2vec/test.py
